Separate base node classes
