+++ src/roman_datamodels/stnode/__init__.py
@@ -1,1 +1,2 @@
+from ._node import *  # noqa: F403
 from ._stnode import *  # noqa: F403
