+++ src/roman_datamodels/stnode/_registry.py
@@ -1,0 +1,4 @@
+OBJECT_NODE_CLASSES_BY_TAG = {}
+LIST_NODE_CLASSES_BY_TAG = {}
+SCALAR_NODE_CLASSES_BY_TAG = {}
+SCALAR_NODE_CLASSES_BY_KEY = {}
