+++ src/roman_datamodels/stnode/_stnode.py
@@ -2,36 +2,29 @@
 Proof of concept of using tags with the data model framework
 """
 
-import datetime
 import importlib.resources
-import re
-import warnings
 from abc import ABCMeta
-from collections import UserList
-from collections.abc import MutableMapping
 
 import asdf
 import asdf.schema as asdfschema
-import asdf.yamlutil as yamlutil
-import jsonschema
-import numpy as np
 import rad.resources
 import yaml
 from asdf.extension import Converter
-from asdf.tags.core import ndarray
-from asdf.util import HashableDict
 from astropy.time import Time
-from astropy.units import Unit  # noqa: F401
 
-from roman_datamodels.validate import ValidationWarning, _check_type, _error_message, will_strict_validate, will_validate
+from ._node import DNode, LNode
+from ._registry import (
+    LIST_NODE_CLASSES_BY_TAG,
+    OBJECT_NODE_CLASSES_BY_TAG,
+    SCALAR_NODE_CLASSES_BY_KEY,
+    SCALAR_NODE_CLASSES_BY_TAG,
+)
 
 __all__ = [
     "WfiMode",
     "NODE_CLASSES",
     "CalLogs",
     "FileDate",
-    "DNode",
-    "LNode",
     "TaggedObjectNode",
     "TaggedListNode",
     "TaggedScalarNode",
@@ -39,237 +32,6 @@ __all__ = [
     "TaggedObjectNodeConverter",
     "TaggedScalarNodeConverter",
 ]
-
-
-validator_callbacks = HashableDict(asdfschema.YAML_VALIDATORS)
-validator_callbacks.update({"type": _check_type})
-
-
-def _value_change(path, value, schema, pass_invalid_values, strict_validation, ctx):
-    """
-    Validate a change in value against a schema.
-    Trap error and return a flag.
-    """
-    try:
-        _check_value(value, schema, ctx)
-        update = True
-
-    except jsonschema.ValidationError as error:
-        update = False
-        errmsg = _error_message(path, error)
-        if pass_invalid_values:
-            update = True
-        if strict_validation:
-            raise jsonschema.ValidationError(errmsg)
-        else:
-            warnings.warn(errmsg, ValidationWarning)
-    return update
-
-
-def _check_value(value, schema, validator_context):
-    """
-    Perform the actual validation.
-    """
-
-    temp_schema = {"$schema": "http://stsci.edu/schemas/asdf-schema/0.1.0/asdf-schema"}
-    temp_schema.update(schema)
-    validator = asdfschema.get_validator(temp_schema, validator_context, validator_callbacks)
-
-    validator.validate(value, _schema=temp_schema)
-    validator_context.close()
-
-
-def _validate(attr, instance, schema, ctx):
-    tagged_tree = yamlutil.custom_tree_to_tagged_tree(instance, ctx)
-    return _value_change(attr, tagged_tree, schema, False, will_strict_validate(), ctx)
-
-
-def _get_schema_for_property(schema, attr):
-    # Check if attr is a property
-    subschema = schema.get("properties", {}).get(attr, None)
-
-    # Check if attr is a pattern property
-    props = schema.get("patternProperties", {})
-    for key, value in props.items():
-        if re.match(key, attr):
-            subschema = value
-            break
-
-    if subschema is not None:
-        return subschema
-    for combiner in ["allOf", "anyOf"]:
-        for subschema in schema.get(combiner, []):
-            subsubschema = _get_schema_for_property(subschema, attr)
-            if subsubschema != {}:
-                return subsubschema
-
-    return {}
-
-
-class DNode(MutableMapping):
-    _tag = None
-    _ctx = None
-
-    def __init__(self, node=None, parent=None, name=None):
-        if node is None:
-            self.__dict__["_data"] = {}
-        elif isinstance(node, dict):
-            self.__dict__["_data"] = node
-        else:
-            raise ValueError("Initializer only accepts dicts")
-        self._x_schema = None
-        self._schema_uri = None
-        self._parent = parent
-        self._name = name
-
-    @property
-    def ctx(self):
-        if self._ctx is None:
-            DNode._ctx = asdf.AsdfFile()
-        return self._ctx
-
-    @staticmethod
-    def _convert_to_scalar(key, value):
-        if key in _SCALAR_NODE_CLASSES_BY_KEY:
-            value = _SCALAR_NODE_CLASSES_BY_KEY[key](value)
-
-        return value
-
-    def __getattr__(self, key):
-        """
-        Permit accessing dict keys as attributes, assuming they are legal Python
-        variable names.
-        """
-        if key.startswith("_"):
-            raise AttributeError(f"No attribute {key}")
-        if key in self._data:
-            value = self._convert_to_scalar(key, self._data[key])
-            if isinstance(value, dict):
-                return DNode(value, parent=self, name=key)
-            elif isinstance(value, list):
-                return LNode(value)
-            else:
-                return value
-        else:
-            raise AttributeError(f"No such attribute ({key}) found in node")
-
-    def __setattr__(self, key, value):
-        """
-        Permit assigning dict keys as attributes.
-        """
-        if key[0] != "_":
-            value = self._convert_to_scalar(key, value)
-            if key in self._data:
-                if will_validate():
-                    schema = _get_schema_for_property(self._schema(), key)
-
-                    if schema == {} or _validate(key, value, schema, self.ctx):
-                        self._data[key] = value
-                self.__dict__["_data"][key] = value
-            else:
-                raise AttributeError(f"No such attribute ({key}) found in node")
-        else:
-            self.__dict__[key] = value
-
-    def to_flat_dict(self, include_arrays=True):
-        """
-        Returns a dictionary of all of the schema items as a flat dictionary.
-
-        Each dictionary key is a dot-separated name.  For example, the
-        schema element ``meta.observation.date`` will end up in the
-        dictionary as::
-
-            { "meta.observation.date": "2012-04-22T03:22:05.432" }
-
-        """
-
-        def convert_val(val):
-            if isinstance(val, datetime.datetime):
-                return val.isoformat()
-            elif isinstance(val, Time):
-                return str(val)
-            return val
-
-        if include_arrays:
-            return {key: convert_val(val) for (key, val) in self.items()}
-        else:
-            return {
-                key: convert_val(val) for (key, val) in self.items() if not isinstance(val, (np.ndarray, ndarray.NDArrayType))
-            }
-
-    def _schema(self):
-        """
-        If not overridden by a subclass, it will search for a schema from
-        the parent class, recursing if necessary until one is found.
-        """
-        if self._x_schema is None:
-            parent_schema = self._parent._schema()
-            # Extract the subschema corresponding to this node.
-            subschema = _get_schema_for_property(parent_schema, self._name)
-            self._x_schema = subschema
-
-        return self._x_schema
-
-    def __asdf_traverse__(self):
-        return dict(self)
-
-    def __len__(self):
-        return len(self._data)
-
-    def __getitem__(self, key):
-        if key in self._data:
-            return self._data[key]
-
-        raise KeyError(f"No such key ({key}) found in node")
-
-    def __setitem__(self, key, value):
-        value = self._convert_to_scalar(key, value)
-        if isinstance(value, dict):
-            for sub_key, sub_value in value.items():
-                value[sub_key] = self._convert_to_scalar(sub_key, sub_value)
-        self._data[key] = value
-
-    def __delitem__(self, key):
-        del self._data[key]
-
-    def __iter__(self):
-        return iter(self._data)
-
-    def copy(self):
-        instance = self.__class__.__new__(self.__class__)
-        instance.__dict__.update(self.__dict__.copy())
-        instance.__dict__["_data"] = self.__dict__["_data"].copy()
-
-        return instance
-
-
-class LNode(UserList):
-    _tag = None
-
-    def __init__(self, node=None):
-        if node is None:
-            self.data = []
-        elif isinstance(node, list):
-            self.data = node
-        elif isinstance(node, self.__class__):
-            self.data = node.data
-        else:
-            raise ValueError("Initializer only accepts lists")
-
-    def __getitem__(self, index):
-        value = self.data[index]
-        if isinstance(value, dict):
-            return DNode(value)
-        elif isinstance(value, list):
-            return LNode(value)
-        else:
-            return value
-
-    def __asdf_traverse__(self):
-        return list(self)
-
-
-_OBJECT_NODE_CLASSES_BY_TAG = {}
 
 
 class TaggedObjectNodeMeta(ABCMeta):
@@ -281,9 +43,9 @@ class TaggedObjectNodeMeta(ABCMeta):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         if self.__name__ != "TaggedObjectNode":
-            if self._tag in _OBJECT_NODE_CLASSES_BY_TAG:
+            if self._tag in OBJECT_NODE_CLASSES_BY_TAG:
                 raise RuntimeError(f"TaggedObjectNode class for tag '{self._tag}' has been defined twice")
-            _OBJECT_NODE_CLASSES_BY_TAG[self._tag] = self
+            OBJECT_NODE_CLASSES_BY_TAG[self._tag] = self
 
 
 class TaggedObjectNode(DNode, metaclass=TaggedObjectNodeMeta):
@@ -309,9 +71,6 @@ class TaggedObjectNode(DNode, metaclass=TaggedObjectNodeMeta):
         return schema
 
 
-_LIST_NODE_CLASSES_BY_TAG = {}
-
-
 class TaggedListNodeMeta(ABCMeta):
     """
     Metaclass for TaggedListNode that maintains a registry
@@ -321,19 +80,15 @@ class TaggedListNodeMeta(ABCMeta):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         if self.__name__ != "TaggedListNode":
-            if self._tag in _LIST_NODE_CLASSES_BY_TAG:
+            if self._tag in LIST_NODE_CLASSES_BY_TAG:
                 raise RuntimeError(f"TaggedListNode class for tag '{self._tag}' has been defined twice")
-            _LIST_NODE_CLASSES_BY_TAG[self._tag] = self
+            LIST_NODE_CLASSES_BY_TAG[self._tag] = self
 
 
 class TaggedListNode(LNode, metaclass=TaggedListNodeMeta):
     @property
     def tag(self):
         return self._tag
-
-
-_SCALAR_NODE_CLASSES_BY_TAG = {}
-_SCALAR_NODE_CLASSES_BY_KEY = {}
 
 
 def _scalar_tag_to_key(tag):
@@ -349,10 +104,10 @@ class TaggedScalarNodeMeta(ABCMeta):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         if self.__name__ != "TaggedScalarNode":
-            if self._tag in _SCALAR_NODE_CLASSES_BY_TAG:
+            if self._tag in SCALAR_NODE_CLASSES_BY_TAG:
                 raise RuntimeError(f"TaggedScalarNode class for tag '{self._tag}' has been defined twice")
-            _SCALAR_NODE_CLASSES_BY_TAG[self._tag] = self
-            _SCALAR_NODE_CLASSES_BY_KEY[_scalar_tag_to_key(self._tag)] = self
+            SCALAR_NODE_CLASSES_BY_TAG[self._tag] = self
+            SCALAR_NODE_CLASSES_BY_KEY[_scalar_tag_to_key(self._tag)] = self
 
 
 class TaggedScalarNode(metaclass=TaggedScalarNodeMeta):
@@ -424,11 +179,11 @@ class TaggedObjectNodeConverter(Converter):
 
     @property
     def tags(self):
-        return list(_OBJECT_NODE_CLASSES_BY_TAG.keys())
+        return list(OBJECT_NODE_CLASSES_BY_TAG.keys())
 
     @property
     def types(self):
-        return list(_OBJECT_NODE_CLASSES_BY_TAG.values())
+        return list(OBJECT_NODE_CLASSES_BY_TAG.values())
 
     def select_tag(self, obj, tags, ctx):
         return obj.tag
@@ -437,7 +192,7 @@ class TaggedObjectNodeConverter(Converter):
         return obj._data
 
     def from_yaml_tree(self, node, tag, ctx):
-        return _OBJECT_NODE_CLASSES_BY_TAG[tag](node)
+        return OBJECT_NODE_CLASSES_BY_TAG[tag](node)
 
 
 class TaggedListNodeConverter(Converter):
@@ -447,11 +202,11 @@ class TaggedListNodeConverter(Converter):
 
     @property
     def tags(self):
-        return list(_LIST_NODE_CLASSES_BY_TAG.keys())
+        return list(LIST_NODE_CLASSES_BY_TAG.keys())
 
     @property
     def types(self):
-        return list(_LIST_NODE_CLASSES_BY_TAG.values())
+        return list(LIST_NODE_CLASSES_BY_TAG.values())
 
     def select_tag(self, obj, tags, ctx):
         return obj.tag
@@ -460,7 +215,7 @@ class TaggedListNodeConverter(Converter):
         return list(obj)
 
     def from_yaml_tree(self, node, tag, ctx):
-        return _LIST_NODE_CLASSES_BY_TAG[tag](node)
+        return LIST_NODE_CLASSES_BY_TAG[tag](node)
 
 
 class TaggedScalarNodeConverter(Converter):
@@ -470,11 +225,11 @@ class TaggedScalarNodeConverter(Converter):
 
     @property
     def tags(self):
-        return list(_SCALAR_NODE_CLASSES_BY_TAG.keys())
+        return list(SCALAR_NODE_CLASSES_BY_TAG.keys())
 
     @property
     def types(self):
-        return list(_SCALAR_NODE_CLASSES_BY_TAG.values())
+        return list(SCALAR_NODE_CLASSES_BY_TAG.values())
 
     def select_tag(self, obj, tags, ctx):
         return obj.tag
@@ -493,7 +248,7 @@ class TaggedScalarNodeConverter(Converter):
             converter = ctx.extension_manager.get_converter_for_type(Time)
             node = converter.from_yaml_tree(node, tag, ctx)
 
-        return _SCALAR_NODE_CLASSES_BY_TAG[tag](node)
+        return SCALAR_NODE_CLASSES_BY_TAG[tag](node)
 
 
 _DATAMODELS_MANIFEST_PATH = importlib.resources.files(rad.resources) / "manifests" / "datamodels-1.0.yaml"
@@ -535,12 +290,12 @@ for tag in _DATAMODELS_MANIFEST["tags"]:
         docstring = tag["description"] + "\n\n"
     docstring = docstring + f"Class generated from tag '{tag['tag_uri']}'"
 
-    if tag["tag_uri"] in _OBJECT_NODE_CLASSES_BY_TAG:
-        _OBJECT_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
-    elif tag["tag_uri"] in _LIST_NODE_CLASSES_BY_TAG:
-        _LIST_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
-    elif tag["tag_uri"] in _SCALAR_NODE_CLASSES_BY_TAG:
-        _SCALAR_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
+    if tag["tag_uri"] in OBJECT_NODE_CLASSES_BY_TAG:
+        OBJECT_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
+    elif tag["tag_uri"] in LIST_NODE_CLASSES_BY_TAG:
+        LIST_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
+    elif tag["tag_uri"] in SCALAR_NODE_CLASSES_BY_TAG:
+        SCALAR_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
     else:
         _class_from_tag(tag, docstring)
 
@@ -548,7 +303,7 @@ for tag in _DATAMODELS_MANIFEST["tags"]:
 # List of node classes made available by this library.  This is part
 # of the public API.
 NODE_CLASSES = (
-    list(_OBJECT_NODE_CLASSES_BY_TAG.values())
-    + list(_LIST_NODE_CLASSES_BY_TAG.values())
-    + list(_SCALAR_NODE_CLASSES_BY_TAG.values())
+    list(OBJECT_NODE_CLASSES_BY_TAG.values())
+    + list(LIST_NODE_CLASSES_BY_TAG.values())
+    + list(SCALAR_NODE_CLASSES_BY_TAG.values())
 )
